requirements.txt
python-dotenv==1.0.1
psycopg2-binary==2.9.7
langchain==0.2.1
langchain-openai==0.1.8
openai>=1.26.0,<2.0.0